+++ virtualizarr/tests/__init__.py
@@ -1,7 +1,38 @@
-import numpy as np
+import importlib
+import itertools
 
-from virtualizarr.manifests import ChunkEntry, ChunkManifest, ManifestArray
-from virtualizarr.zarr import ZArray
+import numpy as np
+import pytest
+from packaging.version import Version
+
+from virtualizarr.manifests import ChunkManifest, ManifestArray
+from virtualizarr.manifests.manifest import join
+from virtualizarr.zarr import ZArray, ceildiv
+
+network = pytest.mark.network
+
+
+def _importorskip(
+    modname: str, minversion: str | None = None
+) -> tuple[bool, pytest.MarkDecorator]:
+    try:
+        mod = importlib.import_module(modname)
+        has = True
+        if minversion is not None:
+            v = getattr(mod, "__version__", "999")
+            if Version(v) < Version(minversion):
+                raise ImportError("Minimum version not satisfied")
+    except ImportError:
+        has = False
+
+    reason = f"requires {modname}"
+    if minversion is not None:
+        reason += f">={minversion}"
+    func = pytest.mark.skipif(not has, reason=reason)
+    return has, func
+
+
+has_s3fs, requires_s3fs = _importorskip("s3fs")
 
 
 def create_manifestarray(
@@ -9,6 +40,8 @@ def create_manifestarray(
 ) -> ManifestArray:
     """
     Create an example ManifestArray with sensible defaults.
+
+    The manifest is populated with a (somewhat) unique path, offset, and length for each key.
     """
 
     zarray = ZArray(
@@ -22,14 +55,38 @@ def create_manifestarray(
         zarr_format=2,
     )
 
-    if shape != ():
-        raise NotImplementedError(
-            "Only generation of array representing a single scalar currently supported"
-        )
-
-    # TODO generalize this
-    chunkmanifest = ChunkManifest(
-        entries={"0": ChunkEntry(path="scalar.nc", offset=6144, length=48)}
+    chunk_grid_shape = tuple(
+        ceildiv(axis_length, chunk_length)
+        for axis_length, chunk_length in zip(shape, chunks)
     )
 
+    if chunk_grid_shape == ():
+        d = {"0": entry_from_chunk_key((0,))}
+    else:
+        # create every possible combination of keys
+        all_possible_combos = itertools.product(
+            *[range(length) for length in chunk_grid_shape]
+        )
+        d = {join(ind): entry_from_chunk_key(ind) for ind in all_possible_combos}
+
+    chunkmanifest = ChunkManifest(entries=d)
+
     return ManifestArray(chunkmanifest=chunkmanifest, zarray=zarray)
+
+
+def entry_from_chunk_key(ind: tuple[int, ...]) -> dict[str, str | int]:
+    """Generate a (somewhat) unique manifest entry from a given chunk key"""
+    entry = {
+        "path": f"file.{str(join(ind))}.nc",
+        "offset": offset_from_chunk_key(ind),
+        "length": length_from_chunk_key(ind),
+    }
+    return entry  # type: ignore[return-value]
+
+
+def offset_from_chunk_key(ind: tuple[int, ...]) -> int:
+    return sum(ind) * 10
+
+
+def length_from_chunk_key(ind: tuple[int, ...]) -> int:
+    return sum(ind) + 5