+++ virtualizarr/tests/test_zarr.py
@@ -3,19 +3,21 @@ import xarray as xr
 import xarray.testing as xrt
 
 from virtualizarr import ManifestArray, open_virtual_dataset
-from virtualizarr.manifests.manifest import ChunkEntry
+from virtualizarr.manifests.manifest import ChunkManifest
 
 
 def test_zarr_v3_roundtrip(tmpdir):
     arr = ManifestArray(
-        chunkmanifest={"0.0": ChunkEntry(path="test.nc", offset=6144, length=48)},
+        chunkmanifest=ChunkManifest(
+            entries={"0.0": dict(path="test.nc", offset=6144, length=48)}
+        ),
         zarray=dict(
             shape=(2, 3),
             dtype=np.dtype("<i8"),
             chunks=(2, 3),
             compressor=None,
             filters=None,
-            fill_value=np.NaN,
+            fill_value=np.nan,
             order="C",
             zarr_format=3,
         ),