+++ docs/releases.rst
@@ -1,10 +1,45 @@
 Release notes
 =============
 
+.. _v0.2:
+
+v0.2 (unreleased)
+-----------------
+
+New Features
+~~~~~~~~~~~~
+
+
+Breaking changes
+~~~~~~~~~~~~~~~~
+
+- Requires numpy 2.0 (for :pull:`107`).
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+
+Deprecations
+~~~~~~~~~~~~
+
+
+Bug fixes
+~~~~~~~~~
+
+
+Documentation
+~~~~~~~~~~~~~
+
+
+Internal Changes
+~~~~~~~~~~~~~~~~
+
+- Refactor `ChunkManifest` class to store chunk references internally using numpy arrays.
+  (:pull:`107`) By `Tom Nicholas <https://github.com/TomNicholas>`_.
+- Mark tests which require network access so that they are only run when `--run-network-tests` is passed a command-line argument to pytest.
+  (:pull:`144`) By `Tom Nicholas <https://github.com/TomNicholas>`_.
+
 .. _v0.1:
 
-v0.1 (unreleased)
------------------
+v0.1 (17th June 2024)
+---------------------
 
 v0.1 is the first release of VirtualiZarr!! It contains functionality for using kerchunk to find byte ranges in netCDF files,
 constructing an xarray.Dataset containing ManifestArray objects, then writing out such a dataset to kerchunk references as either json or parquet.